Create a andiyH package and updated Codes from Andriy: FINRA, Division, OddOrEven
