+++ src/week1/DivisionWithoutOperator.java
@@ -11,42 +11,5 @@ public class DivisionWithoutOperator {
 
     // Solution Below
 
-    public static void main(String[] args) {
-        Scanner input = new Scanner(System.in);
-        System.out.println("Enter your first number: ");
-        int divindend = input.nextInt();
 
-        System.out.println("Enter your second number: ");
-        int divisor = input.nextInt();
-
-        try {
-            divisionWithoutOperator(divindend, divisor);
-        } catch (RuntimeException e) {
-            System.err.println("Divisor can not be 0");
-
-            System.out.println("Try again");
-
-            System.out.println("Enter your first number: ");
-            divindend = input.nextInt();
-
-            System.out.println("Enter your second number: ");
-            divisor = input.nextInt();
-
-            divisionWithoutOperator(divindend, divisor);
-        }
-        input.close();
-    }
-
-
-    public static void divisionWithoutOperator(int dividend, int divisor) {
-        int result = 0;
-        int remainder = dividend % divisor;
-
-            while (dividend >= divisor) {
-                dividend -= divisor;
-                result++;
-            }
-            System.out.println(result + " with reminder " + remainder);
-
-        }
     }
